requirements.txt
python-dotenv

supabase

fastapi

uvicorn
pyjwt